Fix MQTT with Jackson
mqtt-core was based on serde-jackson which may be missing if the consuming project is using jackson.

This change mirrors the fix for problem-json here https://github.com/micronaut-projects/micronaut-problem-json/pull/293

That is, make the serde-api and jackson annotations api dependencies of this project in place of the serde-jackson dependency.

I have an external reproducer, but I cannot get it to fail when included in this project...

Closes #332